more types and non generated keys

+++ src/datastores/my_sql/insert.rs
@@ -15,8 +15,8 @@ pub fn insert_record(
     Ok(())
 }
 
-pub fn last_id(conn: &mut Conn) -> i32 {
-    let id: Result<Vec<i32>, Error> =
+pub fn last_id(conn: &mut Conn) -> i64 {
+    let id: Result<Vec<i64>, Error> =
         conn.query_map("SELECT LAST_INSERT_ID()".to_string(), |id| id);
 
     *id.unwrap().first().unwrap()
